requirements.txt
fastapi==0.109.0          # Web framework
uvicorn[standard]==0.27.0 # ASGI server to run FastAPI
sqlalchemy==2.0.25        # Database ORM (Object Relational Mapper)
psutil==5.9.8             # System metrics collection
python-dotenv==1.0.1      # Load environment variables
pydantic==2.5.3           # Data validation
pydantic-settings==2.1.0  # Settings management